Take out app functions from components
I don't have a clear way of mocking `app` so I want to remove this functionality and pass it in as needed from above, rather than have it baked into the components
